+++ src/hooks/useRegisterForm.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { useState, useEffect } from 'react';
+import { useState } from 'react';
 import { useNavigate } from 'react-router-dom';
 import { useToastStore } from '@/hooks/useToastStore';
 import {
